Refactor FreePort into framework/util
